# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/map/autocomplete": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a list of places that match the query string",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "map"
                ],
                "summary": "Get autocomplete suggestions for places",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Input string to search for",
                        "name": "input",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit the number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Location coordinates (lat,lng)",
                        "name": "location",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Search radius in meters",
                        "name": "radius",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include more compound results",
                        "name": "more_compound",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved autocomplete data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schemas.GoongAutoCompleteResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request query",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to get autocomplete data",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/map/geocode": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves geocode information for a specified latitude and longitude",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "map"
                ],
                "summary": "Get geocode data for a given point",
                "parameters": [
                    {
                        "description": "Geocode request parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.GeoCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved geocode data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schemas.GeoCodeLocationResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to get geocode data",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/map/give-ride": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Receives a list of points and returns a route and polyline encoded string for the driver",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "map"
                ],
                "summary": "Create a route for a driver's give ride",
                "parameters": [
                    {
                        "description": "Give ride request details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.GiveRideRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created route",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schemas.GoongDirectionsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to create route",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/map/hitch-ride": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Receives a list of points and returns a route and polyline encoded string for the passenger",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "map"
                ],
                "summary": "Create a route for a passenger's hitch ride",
                "parameters": [
                    {
                        "description": "Hitch ride request details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.HitchRideRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created route",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schemas.GoongDirectionsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to create route",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "helper.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message_en": {
                    "type": "string"
                },
                "message_vi": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "schemas.GeoCodeLocation": {
            "type": "object",
            "properties": {
                "formatted_address": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "place_id": {
                    "type": "string"
                }
            }
        },
        "schemas.GeoCodeLocationResponse": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.GeoCodeLocation"
                    }
                }
            }
        },
        "schemas.GeoCodeRequest": {
            "type": "object",
            "required": [
                "point"
            ],
            "properties": {
                "point": {
                    "description": "Point for which the geocode is performed",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schemas.Point"
                        }
                    ]
                }
            }
        },
...
        "schemas.GiveRideRequest": {
            "type": "object",
            "required": [
                "place_list"
            ],
            "properties": {
                "place_list": {
                    "description": "Points []Point `json:\"points\" binding:\"required\"` // List of points for the route",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "schemas.GoongAutoCompleteResponse": {
            "type": "object",
            "properties": {
                "executed_time": {
                    "type": "integer"
                },
                "executed_time_all": {
                    "type": "integer"
                },
                "predictions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.Prediction"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "schemas.GoongDirectionsResponse": {
            "type": "object",
            "properties": {
                "geocoded_waypoints": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "geocoder_status": {
                                "type": "string"
                            },
                            "place_id": {
                                "type": "string"
                            }
                        }
                    }
                },
                "routes": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "bounds": {
                                "description": "Empty struct as the bounds are empty in the example",
                                "type": "object"
                            },
                            "legs": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "distance": {
                                            "type": "object",
                                            "properties": {
                                                "text": {
                                                    "type": "string"
                                                },
                                                "value": {
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        "duration": {
                                            "type": "object",
                                            "properties": {
                                                "text": {
                                                    "type": "string"
                                                },
                                                "value": {
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        "end_address": {
                                            "type": "string"
                                        },
                                        "end_location": {
                                            "type": "object",
                                            "properties": {
                                                "lat": {
                                                    "type": "number"
                                                },
                                                "lng": {
                                                    "type": "number"
                                                }
                                            }
                                        },
                                        "start_address": {
                                            "type": "string"
                                        },
                                        "start_location": {
                                            "type": "object",
                                            "properties": {
                                                "lat": {
                                                    "type": "number"
                                                },
                                                "lng": {
                                                    "type": "number"
                                                }
                                            }
                                        },
                                        "steps": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "distance": {
                                                        "type": "object",
                                                        "properties": {
                                                            "text": {
                                                                "type": "string"
                                                            },
                                                            "value": {
                                                                "type": "integer"
                                                            }
                                                        }
                                                    },
                                                    "duration": {
                                                        "type": "object",
                                                        "properties": {
                                                            "text": {
                                                                "type": "string"
                                                            },
                                                            "value": {
                                                                "type": "integer"
                                                            }
                                                        }
                                                    },
                                                    "end_location": {
                                                        "type": "object",
                                                        "properties": {
                                                            "lat": {
                                                                "type": "number"
                                                            },
                                                            "lng": {
                                                                "type": "number"
                                                            }
                                                        }
                                                    },
                                                    "html_instructions": {
                                                        "type": "string"
                                                    },
                                                    "maneuver": {
                                                        "type": "string"
                                                    },
                                                    "polyline": {
                                                        "type": "object",
                                                        "properties": {
                                                            "points": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    },
                                                    "start_location": {
                                                        "type": "object",
                                                        "properties": {
                                                            "lat": {
                                                                "type": "number"
                                                            },
                                                            "lng": {
                                                                "type": "number"
                                                            }
                                                        }
                                                    },
                                                    "travel_mode": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            },
                            "overview_polyline": {
                                "type": "object",
                                "properties": {
                                    "points": {
                                        "type": "string"
                                    }
                                }
                            },
                            "summary": {
                                "type": "string"
                            },
                            "warnings": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "waypoint_order": {
                                "type": "array",
                                "items": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            }
        },
        "schemas.HitchRideRequest": {
            "type": "object",
            "required": [
                "place_list"
            ],
            "properties": {
                "place_list": {
                    "description": "Points []Point `json:\"points\" binding:\"required\"` // List of points for the route",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "schemas.MatchedSubstring": {
            "type": "object",
            "properties": {
                "length": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                }
            }
        },
        "schemas.PlusCode": {
            "type": "object",
            "properties": {
                "compound_code": {
                    "type": "string"
                },
                "global_code": {
                    "type": "string"
                }
            }
        },
        "schemas.Point": {
            "type": "object",
            "properties": {
                "lat": {
                    "description": "Latitude",
                    "type": "number"
                },
                "lng": {
                    "description": "Longitude",
                    "type": "number"
                }
            }
        },
        "schemas.Prediction": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "display_type": {
                    "type": "string"
                },
                "has_children": {
                    "type": "boolean"
                },
                "matched_substrings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.MatchedSubstring"
                    }
                },
                "place_id": {
                    "type": "string"
                },
                "plus_code": {
                    "$ref": "#/definitions/schemas.PlusCode"
                },
                "reference": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                },
                "structured_formatting": {
                    "$ref": "#/definitions/schemas.StructuredFormatting"
                },
                "terms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.Term"
                    }
                }
            }
        },
...
        "schemas.StructuredFormatting": {
            "type": "object",
            "properties": {
                "main_text": {
                    "type": "string"
                },
                "secondary_text": {
                    "type": "string"
                }
            }
        },
        "schemas.Term": {
            "type": "object",
            "properties": {
                "offset": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}